voting algorithm tests added
